+++ lib/index.ts
@@ -1,2 +1,2 @@
-export { schemesFromColor, type Schemes } from './theme';
+export { schemesFromColor, setCSSRules, type Schemes } from './theme';
 export { namedToHex } from './utils';